+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Support loongarch64-unknown-linux-gnu (experimental).
+
 - Performance Improvements.
 
 ## [1.11.2] - 2023-07-18